refactor: move day4 methods to filelib
